+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - 调整反色显示模式色相转换 | Adjusts the inverted display mode Hue transformation.
 - 反向链接面板项应用搜索列表样式 | Backlinks panel items apply the list style of search results.
 - 适配 `v2.4.1` | Adaptive `v2.4.1`.
+- 添加悬浮工具栏 | Add a floating toolbar.
 
 ## v1.3.3/2022-10-12